Full slashmenu impl
