+++ blocks/edit/prose/plugins/slashMenu/slash-menu.css
@@ -5,7 +5,7 @@
   border: 1px solid #ccc;
   border-radius: 4px;
   box-shadow: 0 2px 8px rgb(0 0 0 / 15%);
-  max-height: 300px;
+  max-height: 500px;
   overflow-y: auto;
   z-index: 999999;
   min-width: 150px;
@@ -17,7 +17,7 @@
 
 .slash-menu-items {
   padding: 4px 0;
-  max-height: 300px;
+  max-height: 500px;
   overflow-y: auto;
 }
 
@@ -99,4 +99,20 @@
 
 .slash-menu-icon.open-library {
   background-image: url('/blocks/edit/img/Smock_CCLibrary_18_N.svg');
+}
+
+.slash-menu-icon.lorem-ipsum {
+  background-image: url('/blocks/edit/img/Smock_Rail_18_N.svg');
+}
+
+.slash-menu-item .argument {
+  color: #666;
+  font-style: italic;
+  margin-left: 8px;
+  font-size: 0.9em;
+}
+
+.slash-menu-item .argument.selected {
+  font-style: normal;
+  font-weight: bold;
 }
